+++ packages/catalog/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added the `requiresStringMessageContent` compat option: text-only `messages[].content` is sent as one string for endpoints that reject multi-part arrays ([#12508](https://github.com/can1357/oh-my-pi/pull/12508) by [@ryanskidmore](https://github.com/ryanskidmore)).
+
 ## [18.2.5] - 2026-09-17
 
 ### Added